Django_Backend/requirement.txt
requests
Django==3.2.12
djangorestframework==3.12.4
pyrebase4==4.5.0
Pillow==8.4.0
torch==1.10.0
firebase-admin==5.3.0